Added external google login, joining channel through code, adding/removing channel members
Additionally added cancellation token throughout system

+++ Domain/Entities/Media.cs
@@ -5,6 +5,6 @@ namespace CommunicationsApp.Domain.Entities;
 public class Media : BaseStrongEntity
 {
     public string Filename { get; set; } = null!;
-    public int MediaTypeId { get; set; }
+    public byte MediaTypeId { get; set; }
     public int MessageId { get; set; }
 }
